+++ packages/plugin-e2e/src/fixtures/commands/gotoAppPage.ts
@@ -1,12 +1,8 @@
 import { TestFixture } from '@playwright/test';
-import { PluginFixture, PluginOptions } from '../../api';
-import { PlaywrightCombinedArgs } from '../types';
+import { PlaywrightArgs } from '../../types';
 import { AppPage, GotoAppPageArgs } from '../..';
 
-type GotoAppPageFixture = TestFixture<
-  (args: GotoAppPageArgs) => Promise<AppPage>,
-  PluginFixture & PluginOptions & PlaywrightCombinedArgs
->;
+type GotoAppPageFixture = TestFixture<(args: GotoAppPageArgs) => Promise<AppPage>, PlaywrightArgs>;
 
 export const gotoAppPage: GotoAppPageFixture = async ({ page, selectors, grafanaVersion, request }, use, testInfo) => {
   await use(async ({ pluginId, path }) => {
@@ -15,5 +11,3 @@ export const gotoAppPage: GotoAppPageFixture = async ({ page, selectors, grafana
     return appPage;
   });
 };
-
-export default gotoAppPage;