+++ packages/plugin-e2e/src/fixtures/commands/readProvisionedDashboard.ts
@@ -1,23 +1,19 @@
 import path from 'path';
 import { TestFixture } from '@playwright/test';
 import { promises } from 'fs';
-import { PluginFixture, PluginOptions } from '../../api';
-import { ReadProvisionedDashboardArgs } from '../../types';
-import { PlaywrightCombinedArgs } from '../types';
+import { PlaywrightArgs, ReadProvisionedDashboardArgs } from '../../types';
 
 type ReadProvisionedDashboardFixture = TestFixture<
   <T = any>(args: ReadProvisionedDashboardArgs) => Promise<T>,
-  PluginFixture & PluginOptions & PlaywrightCombinedArgs
+  PlaywrightArgs
 >;
 
 const DASHBOARDS_DIR = 'dashboards';
 
-const readProvisionedDashboard: ReadProvisionedDashboardFixture = async ({ provisioningRootDir }, use) => {
+export const readProvisionedDashboard: ReadProvisionedDashboardFixture = async ({ provisioningRootDir }, use) => {
   await use(async ({ fileName }) => {
     const resolvedPath = path.resolve(path.join(provisioningRootDir, DASHBOARDS_DIR, fileName));
     const contents = await promises.readFile(resolvedPath, 'utf8');
     return JSON.parse(contents);
   });
 };
-
-export default readProvisionedDashboard;