+++ packages/plugin-e2e/src/fixtures/explorePage.ts
@@ -1,14 +1,11 @@
 import { TestFixture } from '@playwright/test';
-import { PluginFixture, PluginOptions } from '../api';
+import { PlaywrightArgs } from '../types';
 import { ExplorePage } from '../models/pages/ExplorePage';
-import { PlaywrightCombinedArgs } from './types';
 
-type ExplorePageFixture = TestFixture<ExplorePage, PluginFixture & PluginOptions & PlaywrightCombinedArgs>;
+type ExplorePageFixture = TestFixture<ExplorePage, PlaywrightArgs>;
 
-const explorePage: ExplorePageFixture = async ({ page, selectors, grafanaVersion, request }, use, testInfo) => {
+export const explorePage: ExplorePageFixture = async ({ page, selectors, grafanaVersion, request }, use, testInfo) => {
   const explorePage = new ExplorePage({ page, selectors, grafanaVersion, request, testInfo });
   await explorePage.goto();
   await use(explorePage);
 };
-
-export default explorePage;