+++ packages/plugin-e2e/src/fixtures/selectors.ts
@@ -1,13 +1,12 @@
 import { TestFixture } from '@playwright/test';
-import { PluginFixture, PluginOptions } from '../api';
+import { PlaywrightArgs } from '../types';
 import { E2ESelectors, resolveSelectors } from '../e2e-selectors';
 import { versionedComponents, versionedPages } from '../e2e-selectors/versioned';
-import { PlaywrightCombinedArgs } from './types';
 import { versionedAPIs } from '../e2e-selectors/versioned/apis';
 
-type SelectorFixture = TestFixture<E2ESelectors, PluginFixture & PluginOptions & PlaywrightCombinedArgs>;
+type SelectorFixture = TestFixture<E2ESelectors, PlaywrightArgs>;
 
-const selectors: SelectorFixture = async ({ grafanaVersion }, use) => {
+export const selectors: SelectorFixture = async ({ grafanaVersion }, use) => {
   const selectors = resolveSelectors(
     {
       components: versionedComponents,
@@ -18,5 +17,3 @@ const selectors: SelectorFixture = async ({ grafanaVersion }, use) => {
   );
   await use(selectors);
 };
-
-export default selectors;