+++ packages/plugin-e2e/src/matchers/toHaveAlert.ts
@@ -1,10 +1,10 @@
 import { expect } from '@playwright/test';
 import { getMessage } from './utils';
-import { GrafanaPage } from '../models';
 import { AlertPageOptions } from '../types';
+import { GrafanaPage } from '../models/pages/GrafanaPage';
 
 export type AlertVariant = 'success' | 'warning' | 'error' | 'info';
-const toHaveAlert = async (grafanaPage: GrafanaPage, severity: AlertVariant, options?: AlertPageOptions) => {
+export const toHaveAlert = async (grafanaPage: GrafanaPage, severity: AlertVariant, options?: AlertPageOptions) => {
   let pass = true;
   let message: any = `An alert of variant ${severity} to be displayed on the page`;
 
@@ -28,5 +28,3 @@ const toHaveAlert = async (grafanaPage: GrafanaPage, severity: AlertVariant, opt
     pass,
   };
 };
-
-export default toHaveAlert;