+++ packages/plugin-e2e/src/options.ts
@@ -1,11 +1,9 @@
 import path = require('path');
 import { Fixtures } from '@playwright/test';
-import { PluginOptions } from '../api';
+import { PluginOptions } from './types';
 
-const options: Fixtures<{}, PluginOptions> = {
+export const options: Fixtures<{}, PluginOptions> = {
   user: [undefined, { option: true, scope: 'worker' }],
   featureToggles: [{}, { option: true, scope: 'worker' }],
   provisioningRootDir: [path.join(process.cwd(), 'provisioning'), { option: true, scope: 'worker' }],
 };
-
-export default options;